requirements.txt
django==2.2.6
pymongo==3.11.0
djangorestframework==3.11.1
dnspython==2.0.0
